% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alecri-package.R
\docType{package}
\name{alecri-package}
\alias{alecri-package}
\title{Several Useful Functions: a Miscellaneous R Package}
\description{
It consists of a collection of functions that may be useful in everyday life
}
\author{
Alessio Crippa, \email{alessio.crippa@ki.se}
}

